+++ Cargo.toml
@@ -37,7 +37,6 @@ lowmemory = []
 [dev-dependencies]
 rand = "0.6"
 rand_core = "0.4"
-serde_test = "1.0"
 
 [dependencies.rand]
 version = "0.6"
@@ -48,3 +47,11 @@ default-features = false
 version = "1.0"
 optional = true
 default-features = false
+
+[dependencies.serde_test]
+version = "1.0"
+optional = true
+
+[[test]]
+name = "serde_structure"
+required-features = [ "serde", "serde_test" ]
