+++ src/key.rs
@@ -817,39 +817,4 @@ mod test {
         assert!(pk3 <= pk1);
         assert!(pk1 >= pk3);
     }
-
-    #[cfg(feature = "serde")]
-    #[test]
-    fn test_signature_serde() {
-        use serde_test::{Configure, Token, assert_tokens};
-        static SK_BYTES: [u8; 32] = [
-            1, 1, 1, 1, 1, 1, 1, 1,
-            0, 1, 2, 3, 4, 5, 6, 7,
-            0xff, 0xff, 0, 0, 0xff, 0xff, 0, 0,
-            99, 99, 99, 99, 99, 99, 99, 99
-        ];
-        static SK_STR: &'static str = "\
-            01010101010101010001020304050607ffff0000ffff00006363636363636363\
-        ";
-        static PK_BYTES: [u8; 33] = [
-            0x02,
-            0x18, 0x84, 0x57, 0x81, 0xf6, 0x31, 0xc4, 0x8f,
-            0x1c, 0x97, 0x09, 0xe2, 0x30, 0x92, 0x06, 0x7d,
-            0x06, 0x83, 0x7f, 0x30, 0xaa, 0x0c, 0xd0, 0x54,
-            0x4a, 0xc8, 0x87, 0xfe, 0x91, 0xdd, 0xd1, 0x66,
-        ];
-        static PK_STR: &'static str = "\
-            0218845781f631c48f1c9709e23092067d06837f30aa0cd0544ac887fe91ddd166\
-        ";
-
-        let s = Secp256k1::new();
-
-        let sk = SecretKey::from_slice(&SK_BYTES).unwrap();
-        let pk = PublicKey::from_secret_key(&s, &sk);
-
-        assert_tokens(&sk.compact(), &[Token::BorrowedBytes(&SK_BYTES[..])]);
-        assert_tokens(&sk.readable(), &[Token::BorrowedStr(SK_STR)]);
-        assert_tokens(&pk.compact(), &[Token::BorrowedBytes(&PK_BYTES[..])]);
-        assert_tokens(&pk.readable(), &[Token::BorrowedStr(PK_STR)]);
-    }
 }
